requirements.txt
valohai-utils
tensorflow==2.6
Pillow
python-multipart
